Extract texture and resource path ownership
