+++ src/core/mesh_builder.py
@@ -1,9 +1,12 @@
 """3DMigoto binary buffer readers and in-memory mesh payload builder."""
 
-import re, struct, os, base64, io, threading, time
-import warnings
-from collections import OrderedDict
+import re, struct, os, base64
 from dataclasses import dataclass
+
+from .resource_paths import safe_resource_path
+from .textures import (_texture_source_uri, _begin_texture_cache,
+                       encode_texture_data_uri, normalize_texture_role,
+                       normalize_texture_transform, texture_key)
 
 # ── Buffer constants ───────────────────────────────────────────────────────────
 POSITION_STRIDE   = 40
@@ -11,132 +14,9 @@ POSITION_OFFSET   = 0
 DEFAULT_UV_OFFSET = 4
 INDEX_SIZE        = 4
 
-# Rendered PNGs are expensive to regenerate during authoring reloads. Keep only
-# the current mod's entries, bounded by rendered byte size. A noisy 2048x2048
-# RGB image can occupy ~12 MiB as PNG bytes, so 256 MiB retains about twenty
-# worst-case textures (and many more typical game textures).
-_TEXTURE_CACHE_LIMIT = 256 * 1024 * 1024
 _MAX_BUFFER_FILE_BYTES = 512 * 1024 * 1024
 _MAX_TOTAL_BUFFER_BYTES = 2 * 1024 * 1024 * 1024
 _MAX_DRAWS = 10_000
-_MAX_IMAGE_PIXELS = 100_000_000
-TEXTURE_ROLES = (
-    "diffuse", "normal_map", "normal_data", "light_map", "material_map")
-TEXTURE_TRANSFORMS = (
-    "passthrough", "normal_xy_reconstruct",
-)
-_texture_cache = OrderedDict()
-_texture_cache_bytes = 0
-_texture_cache_mod = None
-_texture_cache_lock = threading.RLock()
-_texture_profile_hook = None
-
-
-def normalize_texture_role(role=None):
-    """Return the canonical registry role used for one texture instance."""
-    return role if role in TEXTURE_ROLES else "diffuse"
-
-
-def normalize_texture_transform(transform=None):
-    """Return a known image transform, defaulting to raw packed data."""
-    return transform if transform in TEXTURE_TRANSFORMS else "passthrough"
-
-
-def _texture_source_uri(texture_source, path, role, transform):
-    """Call old two-argument and new transform-aware source callbacks."""
-    if texture_source is None:
-        return None
-    try:
-        return texture_source(path, role, transform=transform)
-    except TypeError as first_error:
-        # Direct fixtures and third-party callers historically supplied
-        # ``(path, role)`` callbacks.  Keep those working while the app's
-        # publication registry receives the explicit transform.
-        try:
-            return texture_source(path, role)
-        except TypeError:
-            raise first_error
-
-
-def set_texture_profile_hook(hook):
-    """Install an optional texture-stage callback and return the previous one.
-
-    The normal application leaves this unset, so profiling adds no work to the
-    loading path. Benchmark tooling uses the callback to measure expensive
-    stages without duplicating the renderer's image-processing implementation.
-    """
-    global _texture_profile_hook
-    previous = _texture_profile_hook
-    _texture_profile_hook = hook
-    return previous
-
-
-def _profile_texture(stage, seconds=0.0, **details):
-    hook = _texture_profile_hook
-    if hook is None:
-        return
-    try:
-        hook(stage, seconds, details)
-    except Exception:
-        # Profiling must never change texture loading behavior.
-        pass
-
-
-def _profile_started():
-    return time.perf_counter() if _texture_profile_hook is not None else None
-
-
-def _profile_elapsed(stage, started, **details):
-    if started is not None:
-        _profile_texture(stage, time.perf_counter() - started, **details)
-
-
-def texture_key(relative_path, role=None):
-    """Identify a rendered texture by source path *and* usage role."""
-    role = normalize_texture_role(role)
-    relative_path = str(relative_path or "").replace("\\", "/")
-    return f"{role}::{relative_path}"
-
-
-def split_texture_key(key, default_role=None):
-    """Return ``(role, relative_path)`` for new or legacy texture keys."""
-    value = str(key or "")
-    prefix, separator, relative_path = value.partition("::")
-    if separator and prefix in TEXTURE_ROLES and relative_path:
-        return prefix, relative_path
-    return normalize_texture_role(default_role), value.replace("\\", "/")
-
-
-def normalize_texture_key(key, default_role=None):
-    """Canonicalize a new or legacy key without touching its source path."""
-    role, relative_path = split_texture_key(key, default_role)
-    return texture_key(relative_path, role) if relative_path else None
-
-
-def _begin_texture_cache(mod_dir):
-    global _texture_cache_mod, _texture_cache_bytes
-    root = os.path.normcase(os.path.abspath(mod_dir))
-    with _texture_cache_lock:
-        if root != _texture_cache_mod:
-            _texture_cache.clear()
-            _texture_cache_bytes = 0
-            _texture_cache_mod = root
-
-
-def _cache_texture(key, png):
-    global _texture_cache_bytes
-    size = len(png)
-    if size > _TEXTURE_CACHE_LIMIT:
-        return
-    with _texture_cache_lock:
-        previous = _texture_cache.pop(key, None)
-        if previous is not None:
-            _texture_cache_bytes -= previous[1]
-        _texture_cache[key] = (png, size)
-        _texture_cache_bytes += size
-        while _texture_cache_bytes > _TEXTURE_CACHE_LIMIT:
-            _old_key, (_old_png, old_size) = _texture_cache.popitem(last=False)
-            _texture_cache_bytes -= old_size
 
 
 def _res_get(resources, name):
@@ -150,36 +30,6 @@ def _res_get(resources, name):
     for k, v in resources.items():
         if k.lower() == nl: return v
     return {}
-
-
-_MAX_ESCAPE_DEPTH = 1   # levels above mod_dir a `filename = ..\...` may reach
-
-
-def safe_resource_path(mod_dir, rel):
-    # Resolve a Resource section's `filename = ...` relative to mod_dir.
-
-    if not rel: return None
-    if os.path.isabs(rel) or os.path.splitdrive(rel)[0]:
-        return None
-    root = os.path.abspath(mod_dir)
-    target = os.path.abspath(os.path.join(root, rel))
-    if not _within(target, root):
-        ceiling = root
-        for _ in range(_MAX_ESCAPE_DEPTH):
-            ceiling = os.path.dirname(ceiling)   # stops at the drive root
-        if not _within(target, ceiling):
-            return None
-    return target
-
-
-# Backward-compatible private name for existing callers.  INI diagnostics use
-# the public spelling so resource loading and health checks cannot drift into
-# different path-safety rules.
-_safe_join = safe_resource_path
-
-
-def _within(target, root):
-    return target == root or target.startswith(root + os.sep)
 
 
 # ── Binary readers ─────────────────────────────────────────────────────────────
@@ -355,203 +205,6 @@ class MeshBuildResult:
     textures: dict
     geometry: GeometryBlob | None = None
 
-def _reconstruct_normal_z(img):
-    """Expand a game-style two-channel XY normal into tangent-space RGB.
-
-    ZZMI normal DDS files commonly store X/Y in red/green while blue contains
-    unrelated or placeholder data. Three.js expects a complete XYZ normal.
-    """
-    source = img.convert("RGB").tobytes()
-    result = bytearray(len(source))
-    for offset in range(0, len(source), 3):
-        x = source[offset] / 127.5 - 1.0
-        y = source[offset + 1] / 127.5 - 1.0
-        z = max(0.0, 1.0 - x * x - y * y) ** 0.5
-        result[offset] = source[offset]
-        result[offset + 1] = source[offset + 1]
-        result[offset + 2] = round((z * 0.5 + 0.5) * 255.0)
-    from PIL import Image
-    return Image.frombytes("RGB", img.size, bytes(result))
-
-
-def _apply_texture_transform(img, texture_transform):
-    texture_transform = normalize_texture_transform(texture_transform)
-    if texture_transform == "normal_xy_reconstruct":
-        return _reconstruct_normal_z(img)
-    return img
-
-
-def _render_texture_png(dds_path, max_size=2048, preserve_alpha=False,
-                        texture_role=None, texture_transform="passthrough"):
-    """Decode and explicitly transform an image into PNG bytes."""
-    try:
-        texture_role = normalize_texture_role(texture_role)
-        texture_transform = normalize_texture_transform(texture_transform)
-        stat = os.stat(dds_path)
-        cache_key = (os.path.normcase(os.path.abspath(dds_path)), stat.st_size,
-                     stat.st_mtime_ns, max_size, preserve_alpha, texture_role,
-                     texture_transform)
-        cache_started = _profile_started()
-        with _texture_cache_lock:
-            cached = _texture_cache.pop(cache_key, None)
-            if cached is not None:
-                _texture_cache[cache_key] = cached
-                _profile_elapsed(
-                    "cache_hit", cache_started,
-                    path=cache_key[0], role=texture_role,
-                    transform=texture_transform, bytes=len(cached[0]))
-                return cached[0]
-        _profile_elapsed("cache_miss", cache_started,
-                         path=cache_key[0], role=texture_role,
-                         transform=texture_transform)
-        from PIL import Image
-        Image.MAX_IMAGE_PIXELS = _MAX_IMAGE_PIXELS
-        stage_started = _profile_started()
-        with warnings.catch_warnings():
-            warnings.simplefilter("error", Image.DecompressionBombWarning)
-            img = Image.open(dds_path)
-            img.load()
-        _profile_elapsed("decode", stage_started,
-                         path=cache_key[0], role=texture_role,
-                         transform=texture_transform)
-        # Preserve authored RGBA for packed maps and explicit alpha requests.
-        # Derived normal images intentionally become RGB after their source
-        # channels have been interpreted.
-        packed_passthrough = (
-            texture_transform == "passthrough"
-            and texture_role != "diffuse")
-        keep_source_alpha = preserve_alpha or packed_passthrough
-        stage_started = _profile_started()
-        try:
-            img = img.convert('RGBA' if keep_source_alpha else 'RGB')
-            if preserve_alpha and img.getchannel('A').getextrema()[1] == 0:
-                # Some menu packs deliberately point several slots at an empty
-                # placeholder DDS. Sending it to the browser produces a blank
-                # or black-looking tile; let the menu UI use its cycle glyph.
-                return None
-        finally:
-            _profile_elapsed(
-                "rgb_rgba_conversion", stage_started,
-                path=cache_key[0], role=texture_role,
-                transform=texture_transform)
-        if max(img.size) > max_size:
-            stage_started = _profile_started()
-            img.thumbnail((max_size, max_size), Image.LANCZOS)
-            _profile_elapsed(
-                "resize", stage_started,
-                path=cache_key[0], role=texture_role,
-                transform=texture_transform)
-        if texture_transform != "passthrough":
-            stage_started = _profile_started()
-            try:
-                img = _apply_texture_transform(img, texture_transform)
-            finally:
-                _profile_elapsed(
-                    "normal_z_reconstruction",
-                    stage_started,
-                    path=cache_key[0], role=texture_role,
-                    transform=texture_transform)
-        stage_started = _profile_started()
-        try:
-            buf = io.BytesIO()
-            img.save(buf, format="PNG")
-            png = buf.getvalue()
-        finally:
-            _profile_elapsed(
-                "png_encoding", stage_started,
-                path=cache_key[0], role=texture_role,
-                transform=texture_transform,
-                bytes=len(png) if "png" in locals() else 0)
-        _cache_texture(cache_key, png)
-        _profile_texture(
-            "encoded", 0.0, path=cache_key[0], role=texture_role,
-            transform=texture_transform,
-            bytes=len(png))
-        return png
-    except Exception as e:
-        print(f"  texture skipped: {e}")
-        return None
-
-
-def _encode_texture(dds_path, max_size=2048, preserve_alpha=False,
-                    texture_role=None, texture_transform="passthrough"):
-    """Return the historical base64 data URI compatibility representation."""
-    png = _render_texture_png(dds_path, max_size=max_size,
-                              preserve_alpha=preserve_alpha,
-                              texture_role=texture_role,
-                              texture_transform=texture_transform)
-    if png is None:
-        return None
-    return "data:image/png;base64," + base64.b64encode(png).decode()
-
-
-def encode_texture_file(mod_dir, abs_path, texture_role=None,
-                        texture_source=None, texture_profile=None,
-                        texture_transform=None):
-    """Resolve a picked file into ``{tex_key, file, role, uri}``.
-
-    Without ``texture_source`` the URI is the historical eager data URI. With
-    a callback, the callback publishes the source and supplies its URI.
-
-    The result uses the same role-aware registry shape as the mesh payload, so
-    it can be merged straight into the frontend's shared registry
-    (mesh-factory.js's addTexture). ``abs_path`` must resolve inside
-    ``mod_dir`` -- reuses
-    _safe_join's sandboxing by re-deriving a mod-relative path and rejecting
-    anything that doesn't stay within it, the same constraint `filename =`
-    resolution is already held to.
-    """
-    texture_role = normalize_texture_role(texture_role)
-    if texture_transform is None and (texture_source is None
-                                      or texture_profile is not None):
-        from .texture_profiles import texture_profile_for
-        texture_transform = texture_profile_for(texture_profile).recipe_for(
-            texture_role)
-    if texture_transform is not None:
-        texture_transform = normalize_texture_transform(texture_transform)
-    _begin_texture_cache(mod_dir)
-    try:
-        rel = os.path.relpath(abs_path, mod_dir)
-    except ValueError:
-        return {"error": "Selected file is not inside the mod folder."}
-    resolved = _safe_join(mod_dir, rel)
-    selected = os.path.abspath(abs_path)
-    # Native Windows dialogs may return a path whose drive/folder spelling
-    # differs only by case from the folder picker's spelling. Windows resolves
-    # those to the same file, so compare them using the platform's path rules.
-    if (not resolved
-            or os.path.normcase(os.path.normpath(resolved))
-            != os.path.normcase(os.path.normpath(selected))):
-        return {"error": "Selected file is not inside the mod folder."}
-    if not os.path.isfile(abs_path):
-        return {"error": "Selected file does not exist."}
-    if texture_source is None:
-        uri = _encode_texture(
-            abs_path, texture_role=texture_role,
-            texture_transform=texture_transform)
-    else:
-        uri = _texture_source_uri(
-            texture_source, abs_path, texture_role, texture_transform)
-    if not uri:
-        return {"error": "Could not read this file as an image."}
-    relative_path = rel.replace(os.sep, "/")
-    return {"tex_key": texture_key(relative_path, texture_role),
-            "file": relative_path, "role": texture_role, "uri": uri}
-
-
-def encode_texture_key(mod_dir, key, texture_role=None, texture_source=None,
-                        texture_profile=None, texture_transform=None):
-    """Encode a role-aware registry key, accepting legacy path-only keys."""
-    role, relative_path = split_texture_key(key, texture_role)
-    resolved = _safe_join(mod_dir, relative_path)
-    if not resolved:
-        return {"error": "Selected file is not inside the mod folder."}
-    return encode_texture_file(mod_dir, resolved, role,
-                               texture_source=texture_source,
-                               texture_profile=texture_profile,
-                               texture_transform=texture_transform)
-
-
 # ── In-memory mesh payload builder ────────────────────────────────────────────
 
 def _rel_source(src, mod_dir):
@@ -621,12 +274,12 @@ def _build_shape_buffers(shape_sliders, mod_dir, effective_pos_path, used,
     """Load and prepare dense or sparse shape targets for one draw."""
     shape_buffers = []
     for shape in shape_sliders or []:
-        shape_base_path = _safe_join(mod_dir, shape["base_file"])
+        shape_base_path = safe_resource_path(mod_dir, shape["base_file"])
         if os.path.normcase(os.path.normpath(shape_base_path or "")) != \
                 os.path.normcase(os.path.normpath(effective_pos_path)):
             continue
         if shape.get("shape_id") is not None:
-            paths = tuple(_safe_join(mod_dir, shape[k]) for k in
+            paths = tuple(safe_resource_path(mod_dir, shape[k]) for k in
                           ("offset_file", "vertex_id_file", "vertex_offset_file"))
             if not all(path and os.path.exists(path) for path in paths):
                 continue
@@ -660,7 +313,7 @@ def _build_shape_buffers(shape_sliders, mod_dir, effective_pos_path, used,
             shape_buffers.append((shape, sparse_shape_cache[cache_key],
                                   bytearray(len(used) * 12), True))
         else:
-            target_path = _safe_join(mod_dir, shape["target_file"])
+            target_path = safe_resource_path(mod_dir, shape["target_file"])
             if not target_path or not os.path.exists(target_path):
                 continue
             if target_path not in raw_buf_cache:
@@ -668,7 +321,7 @@ def _build_shape_buffers(shape_sliders, mod_dir, effective_pos_path, used,
             low_data = None
             low_bytes = None
             if shape.get("low_file"):
-                low_path = _safe_join(mod_dir, shape["low_file"])
+                low_path = safe_resource_path(mod_dir, shape["low_file"])
                 if not low_path or not os.path.exists(low_path):
                     continue
                 if low_path not in raw_buf_cache:
@@ -750,9 +403,9 @@ def build_mesh_result(groups, mod_dir, max_draws=0, geometry=None,
         raise ValueError(f"Mod has too many draws ({draw_total:,}; limit {_MAX_DRAWS:,}).")
 
     for grp in groups:
-        pos_path  = _safe_join(mod_dir, grp["position_file"])
-        tc_path   = _safe_join(mod_dir, grp["texcoord_file"])
-        ib_path   = _safe_join(mod_dir, grp["ib_file"])
+        pos_path  = safe_resource_path(mod_dir, grp["position_file"])
+        tc_path   = safe_resource_path(mod_dir, grp["texcoord_file"])
+        ib_path   = safe_resource_path(mod_dir, grp["ib_file"])
         tc_stride = grp["texcoord_stride"]
         pos_stride = grp.get("position_stride", POSITION_STRIDE)
         index_size = grp.get("index_size", INDEX_SIZE)
@@ -810,7 +463,7 @@ def build_mesh_result(groups, mod_dir, max_draws=0, geometry=None,
             key = _tex_key(dds_path, texture_role, texture_transform)
             if key and key not in tex_uris:
                 if texture_source is None:
-                    value = _encode_texture(
+                    value = encode_texture_data_uri(
                         dds_path, texture_role=texture_role,
                         texture_transform=texture_transform)
                 else:
@@ -827,7 +480,7 @@ def build_mesh_result(groups, mod_dir, max_draws=0, geometry=None,
         # that's also someone's active default doesn't get encoded twice.
         texture_options = []
         for pool_entry in grp.get("diffuse_pool_files") or []:
-            key = _tex_key(_safe_join(mod_dir, pool_entry["file"]))
+            key = _tex_key(safe_resource_path(mod_dir, pool_entry["file"]))
             if key:
                 res_name = pool_entry["res"]
                 label = res_name[8:] if res_name.startswith("Resource") else res_name
@@ -839,7 +492,7 @@ def build_mesh_result(groups, mod_dir, max_draws=0, geometry=None,
             lbl = draw["label"]
             draw_ib_path = ib_path
             if draw.get("ib_file"):
-                draw_ib_path = _safe_join(mod_dir, draw["ib_file"])
+                draw_ib_path = safe_resource_path(mod_dir, draw["ib_file"])
                 if not draw_ib_path or not os.path.exists(draw_ib_path):
                     continue
                 if draw_ib_path not in ib_cache:
@@ -863,8 +516,8 @@ def build_mesh_result(groups, mod_dir, max_draws=0, geometry=None,
             draw_buffers = buffers
             effective_pos_path = pos_path
             if draw.get("position_file") and draw.get("texcoord_file"):
-                draw_pos_path = _safe_join(mod_dir, draw["position_file"])
-                draw_tc_path  = _safe_join(mod_dir, draw["texcoord_file"])
+                draw_pos_path = safe_resource_path(mod_dir, draw["position_file"])
+                draw_tc_path  = safe_resource_path(mod_dir, draw["texcoord_file"])
                 if not (draw_pos_path and draw_tc_path
                         and os.path.exists(draw_pos_path) and os.path.exists(draw_tc_path)):
                     continue
@@ -921,7 +574,7 @@ def build_mesh_result(groups, mod_dir, max_draws=0, geometry=None,
                 struct.pack_into("<I", idx_bytes, out_i * 4, remap[value])
 
             default_key = _ensure_texture(
-                _safe_join(mod_dir, draw.get("texture_default_file")))
+                safe_resource_path(mod_dir, draw.get("texture_default_file")))
             entry: dict = {
                 "pos": _geometry_ref(pos_bytes, geometry),
                 "idx": _geometry_ref(idx_bytes, geometry),
@@ -932,14 +585,14 @@ def build_mesh_result(groups, mod_dir, max_draws=0, geometry=None,
             # NormalMap is a user-facing authored role, but its transport is
             # profile-owned.  WuWa publishes the intact packed source as
             # normal_data; Genshin/ZZZ retain the derived normal_map path.
-            normal_path = _safe_join(
+            normal_path = safe_resource_path(
                 mod_dir, draw.get("normal_map_default_file"))
             normal_role = texture_profile.normal_transport_role
             normal_key = _ensure_texture(normal_path, normal_role)
             if normal_key:
                 entry[f"{normal_role}_key"] = normal_key
             for channel in ("light_map", "material_map"):
-                key = _ensure_texture(_safe_join(
+                key = _ensure_texture(safe_resource_path(
                     mod_dir, draw.get(f"{channel}_default_file")), channel)
                 if key:
                     entry[f"{channel}_key"] = key
@@ -981,7 +634,7 @@ def build_mesh_result(groups, mod_dir, max_draws=0, geometry=None,
             if texture_rules:
                 variants = []
                 for v in texture_rules:
-                    key = _ensure_texture(_safe_join(mod_dir, v["file"]))
+                    key = _ensure_texture(safe_resource_path(mod_dir, v["file"]))
                     if key:
                         # Auxiliary assignments after a conditional diffuse
                         # branch belong to every branch reaching this draw,
@@ -995,7 +648,7 @@ def build_mesh_result(groups, mod_dir, max_draws=0, geometry=None,
                 variants = []
                 for variant in rules:
                     key = _ensure_texture(
-                        _safe_join(mod_dir, variant["file"]), channel)
+                        safe_resource_path(mod_dir, variant["file"]), channel)
                     if key:
                         variants.append({
                             "conditions": variant["conditions"],
@@ -1007,7 +660,7 @@ def build_mesh_result(groups, mod_dir, max_draws=0, geometry=None,
             normal_variants = []
             for variant in normal_rules:
                 key = _ensure_texture(
-                    _safe_join(mod_dir, variant["file"]), normal_role)
+                    safe_resource_path(mod_dir, variant["file"]), normal_role)
                 if key:
                     normal_variants.append({
                         "conditions": variant["conditions"],
